Move manifest validation logic to separate file
